refactor: separate board, piece type and constants; update imports
