+++ src/lib/constants/board.ts
@@ -1,0 +1,5 @@
+// BOARD CONSTANTS
+
+// board dimensions
+export const BOARD_ROWS = 5;
+export const BOARD_COLS = 11;
