+++ src/types/board-and-piece-types.ts
@@ -1,0 +1,17 @@
+// BOARD, PIECE TYPES
+
+export type Board = number[][];
+
+export type Coordinate = [number, number];
+
+export interface Piece {
+    letter: string;
+    color: string;
+    base: number[][];
+    variations: Coordinate[][];
+    url?: string;
+}
+
+export interface PieceCollection {
+    [key: number]: Piece;
+}
