--- src/types/board.ts
@@ -1,9 +1,0 @@
-// BOARD TYPES AND CONSTANTS
-
-// types
-export type Board = number[][];
-export type Coordinate = [number, number];
-
-// board dimensions
-export const BOARD_ROWS = 5;
-export const BOARD_COLS = 11;
